+++ src/utils/account/account.ts
@@ -1,10 +1,10 @@
 import { Address, log } from '@graphprotocol/graph-ts';
-
 import {
   Account,
-} from '../../generated/schema';
+} from '../../../generated/schema';
+import * as vaultPosition from './vault-position'
 
-export function getOrCreateAccount(address: Address): Account {
+export function getOrCreate(address: Address): Account {
   log.debug('[Account] Get or create account', [])
   let id = address.toHexString();
   let account = Account.load(id);
@@ -16,4 +16,3 @@ export function getOrCreateAccount(address: Address): Account {
 
   return account as Account;
 }
-
