+++ src/utils/account/vault-position-update.ts
@@ -1,0 +1,6 @@
+import { AccountVaultPositionUpdate } from "../../../generated/schema";
+
+
+export function getOrCreate(): AccountVaultPositionUpdate {
+  return null
+}
