+++ src/utils/vault/vault.ts
@@ -1,0 +1,216 @@
+import { Address, ethereum, BigInt, log, Bytes  } from "@graphprotocol/graph-ts";
+import {
+  Vault,
+} from "../../../generated/schema";
+
+import { Vault as VaultContract } from '../../../generated/Registry/Vault';
+import { Vault as VaultTemplate } from "../../../generated/templates";
+import { BIGINT_ZERO } from '../constants';
+import { getOrCreateToken } from '../token';
+import { createStrategy } from "../strategy";
+import * as depositLibrary from '../deposit'
+import * as accountLibrary from '../account/account'
+import * as accountVaultPositionLibrary from '../account/vault-position'
+import * as vaultUpdateLibrary from './vault-update'
+
+const createNewVaultFromAddress = (vaultAddress: Address): Vault => {
+  let id = vaultAddress.toHexString();
+  let vaultEntity = new Vault(id);
+  let vaultContract = VaultContract.bind(vaultAddress);
+
+  let token = getOrCreateToken(vaultContract.token());
+  let shareToken = getOrCreateToken(vaultAddress);
+
+  // TODO Create transaction vaultEntity.transaction = transactionId
+  vaultEntity.transaction = "0";
+  vaultEntity.token = token.id;
+  vaultEntity.shareToken = shareToken.id;
+
+  // vault fields
+  vaultEntity.activation = vaultContract.activation();
+  vaultEntity.apiVersion = vaultContract.apiVersion();
+
+  // NOTE: derived
+  // vaultEntity.strategies = [];
+
+  // empty at creation
+  vaultEntity.tags = [];
+  vaultEntity.balanceTokens = BIGINT_ZERO;
+  vaultEntity.balanceTokensIdle = BIGINT_ZERO;
+  vaultEntity.balanceTokensInvested = BIGINT_ZERO;
+
+  vaultEntity.tokensDepositLimit = BIGINT_ZERO;
+  vaultEntity.sharesSupply = BIGINT_ZERO;
+  vaultEntity.managementFeeBps = 0;
+  vaultEntity.performanceFeeBps = 0;
+
+  // vaultEntity.tokensDepositLimit = vaultContract.depositLimit();
+  // vaultEntity.sharesSupply = vaultContract.totalSupply();
+  // vaultEntity.managementFeeBps = vaultContract.managementFee().toI32();
+  // vaultEntity.performanceFeeBps = vaultContract.performanceFee().toI32();
+
+
+  return vaultEntity;
+}
+
+export function getOrCreate(vaultAddress: Address, createTemplate: boolean): Vault {
+  let id = vaultAddress.toHexString();
+  let vault = Vault.load(id);
+
+  if (vault == null) {
+    vault = createNewVaultFromAddress(vaultAddress)
+
+    if(createTemplate) {
+      VaultTemplate.create(vaultAddress);
+    }
+  }
+
+  return vault as Vault;
+}
+
+export function create(
+  transactionId: string,
+  vault: Address,
+  classification: string,
+  apiVersion: string,
+  deploymentId: BigInt,
+  createTemplate: boolean,
+  event: ethereum.Event
+): Vault {
+  log.info('[Vault] Create vault', [])
+  let id = vault.toHexString()
+  let vaultEntity = Vault.load(id)
+  if(vaultEntity == null) {
+    vaultEntity = createNewVaultFromAddress(vault); 
+    vaultEntity.transaction = transactionId
+    vaultEntity.classification = classification
+    // vaultEntity.deploymentId = deploymentId
+    vaultEntity.apiVersion = apiVersion
+    if(createTemplate) {
+      VaultTemplate.create(vault);
+    }
+  } else {
+    // NOTE: vault is experimental but being endorsed
+    if (vaultEntity.classification !== classification) {
+      vaultEntity.classification = classification;
+    }
+  }
+  // vaultEntity.blockNumber = event.block.number
+  // vaultEntity.timestamp = getTimestampInMillis(event)
+  vaultEntity.save()
+  return vaultEntity as Vault
+}
+  
+// TODO: implement this
+export function release(
+  vault: Address,
+  apiVersion: string,
+  releaseId: BigInt,
+  event: ethereum.Event
+): Vault | null {
+  let id = vault.toHexString()
+  let entity = Vault.load(id)
+  if(entity !== null) {
+    // TODO: implement this
+    // entity.status = 'Released'
+    // entity.apiVersion = apiVersion
+    // entity.deploymentId = deploymentId
+    // entity.blockNumber = event.block.number
+    // entity.timestamp = getTimestampInMillis(event)
+    // entity.save()
+  }
+  return entity
+}
+
+export function addStrategy(
+  transactionId: string,
+  vaultAddress: Address,
+  strategy: Address,
+  debtLimit: BigInt,
+  performanceFee: BigInt,
+  rateLimit: BigInt,
+  event: ethereum.Event,
+): void {
+  let id = vaultAddress.toHexString()
+  let entity = Vault.load(id)
+  if(entity !== null) {
+    let newStrategy = createStrategy(
+      transactionId,
+      strategy,
+      vaultAddress,
+      debtLimit,
+      rateLimit,
+      performanceFee,
+      event
+    )
+    let strategies = entity.strategies
+    strategies.push(newStrategy.id)
+    entity.strategies = strategies
+    entity.save()
+  }
+}
+
+export function tag(
+  vault: Address,
+  tag: string
+): Vault {
+  let id = vault.toHexString()
+  log.info("Processing tag for vault address: {}", [id])
+  let entity = Vault.load(id)
+  if(entity == null) {
+  } else {
+    entity.tags = tag.split(",")
+    entity.save()
+  }
+  return entity as Vault
+}
+
+export function deposit(
+  transactionHash: Bytes,
+  transactionIndex: BigInt,
+  from: Address,
+  to: Address,
+  inputAmount: BigInt,
+  totalAssets: BigInt,
+  totalSupply: BigInt,
+  pricePerShare: BigInt
+): void {
+  let account = accountLibrary.getOrCreate(from)
+  let vault = getOrCreate(to, false)
+  let sharesMinted = totalAssets.equals(BIGINT_ZERO)
+    ? inputAmount
+    : inputAmount.times(totalSupply).div(totalAssets)
+    
+  let vaultPositionResponse = accountVaultPositionLibrary.deposit(
+    account,
+    vault,
+    transactionHash.toHexString(),
+    inputAmount,
+    sharesMinted
+  )
+  
+  // let deposit = depositLibrary.getOrCreate(
+  //   transactionHash,
+  //   transactionIndex,
+  //   account,
+  //   vault,
+  //   inputAmount,
+  //   sharesMinted
+  // )
+  // let vaultUpdate = vaultUpdateLibrary.getOrCreate(
+  //   vault,
+  //   transactionHash,
+  //   transactionIndex,
+  //   inputAmount,
+  //   BIGINT_ZERO,
+  //   sharesMinted,
+  //   BIGINT_ZERO,
+  //   pricePerShare
+  // )
+
+  // vault.historicalUpdates.push(vaultUpdate.id);
+  // vault.latestUpdate = vaultUpdate.id
+  // create vault update
+  // modift internal data
+  // save
+}
