+++ subgraph.template.yaml
@@ -12,7 +12,7 @@ dataSources:
       abi: Registry
     mapping:
       kind: ethereum/events
-      apiVersion: 0.0.3
+      apiVersion: 0.0.4
       language: wasm/assemblyscript
       entities:
         - Vault
@@ -44,7 +44,7 @@ templates:
       abi: Vault
     mapping:
       kind: ethereum/events
-      apiVersion: 0.0.1
+      apiVersion: 0.0.4
       language: wasm/assemblyscript
       file: ./src/mappings/vaultMappings.ts
       entities:
@@ -77,5 +77,11 @@ templates:
           handler: handleDepositWithAmount
         - function: deposit(uint256,address)
           handler: handleDepositWithAmountAndRecipient
-      #   - function: withdraw(uint256)
-      #     handler: handleWithdrawal
+        - function: withdraw()
+          handler: handleWithdraw
+        - function: withdraw(uint256)
+          handler: handleWithdrawWithShares
+        - function: withdraw(uint256,address)
+          handler: handleWithdrawWithSharesAndRecipient
+        # - function: withdraw(uint256,address,uint256)
+        #   handler: handleWithdrawWithSharesRecipientAndMaxLoss
